+++ packages/ui/src/index.ts
@@ -1,6 +1,0 @@
-export * from "@fasti/tokens";
-
-export interface ComponentProps {
-  className?: string;
-  children?: any;
-}